oh-tab-9mc: modularize extension event handlers
